# Conflicts:
#	src/main/java/kernbeisser/CustomComponents/ShoppingTable/ShoppingCartModel.java
#	src/main/java/kernbeisser/CustomComponents/ShoppingTable/ShoppingCartView.java

+++ src/main/java/kernbeisser/CustomComponents/ShoppingTable/ShoppingCartModel.java
@@ -8,11 +8,11 @@ import java.util.HashMap;
 
 public class ShoppingCartModel implements Model {
     private final double userValue;
-    private final int userSurcharge;
+    private final double userSurcharge;
     private HashMap<ShoppingItem,ShoppingItem> shoppingItems = new HashMap<>();
 
 
-    ShoppingCartModel(double userValue, int userSurcharge) {
+    ShoppingCartModel(double userValue, double userSurcharge) {
         this.userValue = userValue;
         this.userSurcharge = userSurcharge;
     }
@@ -34,7 +34,7 @@ public class ShoppingCartModel implements Model {
         return shoppingItems.values();
     }
 
-    public int getUserSurcharge() {
+    public double getUserSurcharge() {
         return userSurcharge;
     }
 }
